Rebase fixes for x64
  - Fixed a few merge issues
  - Fixed the update warnings
  - Retested, and everything bar the about box appears to be working
     - there is on-going work on about box, so not going to put work
       in now to fix it

+++ PowerEditor/src/Parameters.cpp
@@ -2125,7 +2125,7 @@ int NppParameters::addUserLangToEnd(const UserLangContainer & userLang, const TC
 	newLangContainer->setName(newName);
 	_userLangArray.push_back(newLangContainer);
 
-	return _userLangArray.size()-1;
+	return static_cast<int>(_userLangArray.size())-1;
 }
 
 void NppParameters::removeUserLang(int index)
